+++ src/invader.cmake
@@ -75,9 +75,6 @@ set(INVADER_SOURCE_FILES
     src/bitmap/sprite.cpp
     src/error_handler/error_handler.cpp
     src/model/jms.cpp
-    src/script/compiler.cpp
-    src/script/script_tree.cpp
-    src/script/tokenizer.cpp
     src/compress/compression.cpp
     src/tag/hek/header.cpp
     src/tag/hek/class/bitmap.cpp